+++ aspire_dash/__init__.py
@@ -74,7 +74,7 @@ import os
 import shutil
 import dash_bootstrap_components as dbc
 
-__version__ = "0.79.0"
+__version__ = "0.80.0"
 
 
 def normalised_path(pathname: str | None) -> str: